requirements.txt
twitchAPI
python-dotenv
SQLAlchemy
